+++ packages/expo-file-system/src/FileSystem.ts
@@ -12,6 +12,13 @@ export class Paths extends PathUtilities {
   }
 
   /**
+   * A property containing the bundle directory – the directory where assets bundled with the application are stored.
+   */
+  static get bundle() {
+    return new Directory(ExpoFileSystem.bundleDirectory);
+  }
+
+  /**
    * A property containing the document directory – a place to store files that are safe from being deleted by the system.
    */
   static get document() {