+++ packages/expo-router/src/__tests__/initialRouteName.test.ios.tsx
@@ -1,3 +1,4 @@
+import { screen, act } from '@testing-library/react-native';
 import React from 'react';
 import { Text } from 'react-native';
 
@@ -5,7 +6,7 @@ import { store } from '../global-state/router-store';
 import { useLocalSearchParams } from '../hooks';
 import { router } from '../imperative-api';
 import Stack from '../layouts/Stack';
-import { act, renderRouter, screen } from '../testing-library';
+import { renderRouter } from '../testing-library';
 
 /**
  * anchor sets the "default" screen for a navigator, with the functionality changing per navigator