+++ packages/expo-router/src/__tests__/smoke.test.ios.tsx
@@ -1,3 +1,4 @@
+import { act, screen } from '@testing-library/react-native';
 import React, { Text } from 'react-native';
 
 import { Slot, router, useGlobalSearchParams, usePathname } from '../exports';
@@ -5,7 +6,7 @@ import { Drawer } from '../layouts/Drawer';
 import { Stack } from '../layouts/Stack';
 import { Tabs } from '../layouts/Tabs';
 import { Redirect } from '../link/Link';
-import { act, renderRouter, screen } from '../testing-library';
+import { renderRouter } from '../testing-library';
 
 it('404', () => {
   renderRouter(