+++ packages/expo-sqlite/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Updated API doc for `loadExtensionAsync()` and `loadExtensionSync()`. ([#38905](https://github.com/expo/expo/pull/38905) by [@kudo](https://github.com/kudo))
+
 ## 16.0.1 — 2025-08-15
 
 _This version does not introduce any user-facing changes._