+++ packages/expo-sqlite/CHANGELOG.md
@@ -10,12 +10,17 @@
 
 ### 💡 Others
 
+## 16.0.1 — 2025-08-15
+
+_This version does not introduce any user-facing changes._
+
 ## 16.0.0 — 2025-08-13
 
 ### 🎉 New features
 
 - [iOS] Support Apple TV. ([#38475](https://github.com/expo/expo/pull/38475) by [@douglowder](https://github.com/douglowder))
 - Added `loadExtensionAsync` / `loadExtensionSync` APIs and pre-bundled [sqlite-vec](https://github.com/asg017/sqlite-vec) extension. ([#38693](https://github.com/expo/expo/pull/38693) by [@kudo](https://github.com/kudo))
+- Added `expo-sqlite/localStorage/install` for [`localStorage`](https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage) drop-in replacement API. ([#38699](https://github.com/expo/expo/pull/38699) by [@kudo](https://github.com/kudo))
 
 ### 🐛 Bug fixes